pip_requirements.txt
python-libsbml
networkx==2.3
numpy
pandas
openbabel
timeout-decorator
cython
biopython==1.77
equilibrator-api
equilibrator-pathway
objsize
shared_memory_dict
graphviz
pydotplus
lxml

